requirements.txt
transformers
accelerate
evaluate
timm
python-dotenv
huggingface_hub
xformers
diffusers
datasets[vision]